refactored coin info Page a bit
